requirements.txt
asgiref==3.8.1
cpf-cnpj-validate==1.4
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
drf-yasg==1.21.8
gunicorn==20.1.0
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.5.2
uritemplate==4.1.1
